requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
flask==1.0.2
idna==2.7
numexpr
numpy
pandas
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.0
six==1.11.0
tables
urllib3==1.24
pymongo==3.7.2
nose